move the rest of the implementation to ADF
