+++ demo-shell/src/app/components/search/search-result.component.html
@@ -1,19 +1,15 @@
 <div class="adf-search-results__facets">
     <adf-selected-facets
-        [facetQueries]="selectedFacetQueries"
-        [facetBuckets]="selectedBuckets"
-        (removeFacetQuery)="onRemoveFacetQuery($event)"
-        (removeFacetBucket)="onRemoveSelectedBucket($event)">
+        [facetQueries]="settings.selectedFacetQueries"
+        [facetBuckets]="settings.selectedBuckets"
+        (removeFacetQuery)="settings.unselectFacetQuery($event)"
+        (removeFacetBucket)="settings.unselectFacetBucket($event)">
     </adf-selected-facets>
 </div>
 
 <div class="adf-search-results">
     <adf-search-settings #settings
-        [queryBuilder]="queryBuilder"
-        [responseFacetQueries]="responseFacetQueries"
-        [responseFacetFields]="responseFacetFields"
-        (facetQueryToggle)="onFacetQueryToggle($event)"
-        (facetFieldToggle)="onFacetToggle($event)">
+        [queryBuilder]="queryBuilder">
     </adf-search-settings>
 
     <div class="adf-search-results__content">
